+++ changelog.rst
@@ -4,8 +4,7 @@ Changelog
 Release 1.0.19
 --------------
 
-- Enhancement: Added fixed version of 'Europe/Moscow' timezone
-- Enhancement: Added 'CEST' timezone
+- Enhancement: Added :code:`rmr.utils.datetime.fromtimestamp()`
 
 Release 1.0.18
 --------------